Specify bash in the shebang
This file is written for bash and contains several bashisms so it should explicitly use bash and not whatever shell is defined in the users env. This script doesn't run properly with either zsh which is my the shell defined in my env, or dash which is sh is symlinked to in ubuntu.